pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sast-regression-framework"
version = "0.1.0"
description = "Automated regression testing framework for SAST tools in CI/CD"
readme = "README.md"
authors = [
    {name = "SAST Framework Team", email = "team@sast-framework.example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0",
    "pydantic>=2.0",
    "docker>=6.0",
    "jsonschema>=4.0",
    "python-dateutil>=2.8",
    "colorlog>=6.0",
    "typer>=0.9",
    "rich>=13.0",
]

[project.scripts]
sast-regression = "framework.main:main"

[tool.setuptools.packages.find]
where = ["src"]